% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/charSummary.R
\name{charSummary}
\alias{charSummary}
\title{Automatically detects character/factor variables and gives a comprehensive summary}
\usage{
charSummary(df)
}
\arguments{
\item{df}{name of your data frame}
}
\value{
Returns the summary data frame
}
\description{
Automatically detects character/factor variables and gives a comprehensive summary
}
\examples{
data(iris)
charSummary(iris)
}

