% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numSummary.R
\name{numSummary}
\alias{numSummary}
\title{Automatically detects numeric variables and gives a comprehensive summary}
\usage{
numSummary(df)
}
\arguments{
\item{df-}{name of your data frame}
}
\value{
Returns the summary data frame
}
\description{
Automatically detects numeric variables and gives a comprehensive summary
}
\examples{
data(iris)
numSummary(iris)
}

